Removed GetGroupIDByName. Corrections AddGroupMembers and README
