+++ src/main/groovy/org/craftercms/cli/Main.groovy
@@ -17,22 +17,21 @@
 package org.craftercms.cli
 
 import org.craftercms.cli.commands.AddEnvironment
+import org.craftercms.cli.commands.group.AddGroupMembers
 import org.craftercms.cli.commands.group.CreateGroup
-import org.craftercms.cli.commands.group.GetGroupIDByName
 import org.craftercms.cli.commands.group.ListGroups
 import org.craftercms.cli.commands.marketplace.CopyPlugin
 import org.craftercms.cli.commands.security.CreateAccessToken
 import org.craftercms.cli.commands.site.*
 import org.craftercms.cli.commands.user.CreateUser
 import org.craftercms.cli.commands.user.ListUsers
-import org.craftercms.cli.commands.group.AddGroupMembers
 import picocli.CommandLine
 
 @CommandLine.Command(
         name = 'crafter-cli', usageHelpAutoWidth = true,
         versionProvider = VersionProvider.class, mixinStandardHelpOptions = true,
         subcommands = [CommandLine.HelpCommand, AddEnvironment, AddRemote, CreateSite, ListRemotes, SyncFrom, SyncTo,
-                ListSites, CopyPlugin, CreateUser, ListUsers, CreateAccessToken, PublishContent, CreateGroup, ListGroups, AddGroupMembers, GetGroupIDByName])
+                ListSites, CopyPlugin, CreateUser, ListUsers, CreateAccessToken, PublishContent, CreateGroup, ListGroups, AddGroupMembers])
 class Main {
 
     static def main(args) {
